Fixed 2D detection not working due to being checked by the debug bool Added full custom inspector for all nodes. Looks nice and shiny now. Renamed clutter menu to clutter editor, due to the changed nature of the script

+++ ClutterProj/Assets/ClutterBug/3DNode/Clutter.cs
@@ -26,7 +26,7 @@ public class Clutter : MonoBehaviour
 
     [Tooltip("If the collider's angle is less than this value, the clutter wont spawn.")]
     [Range(0, 89)]
-    public int angleLimit = 45;
+    public float angleLimit = 45;
 
     [Tooltip("Randomised rotation value between a Min and Max. Is overriden by rotation override.")]
     [Header("Randomise Rotation")]
@@ -265,7 +265,7 @@ public class Clutter : MonoBehaviour
     }
 
     public Vector3 SetScale(Transform go2)
-    { 
+    {
         Vector3 toReturn = new Vector3(0, 0, 0);
 
         if (scaleOverride == Vector3.zero && randomScale != Vector2.zero)
